+++ src/scene_objects/point_light.hh
@@ -6,9 +6,9 @@ class PointLight : public LightSource
 {
 public:
     Point3 center;
-    float radius;
+    double radius;
 
-    PointLight(Point3 center, float intensity, float radius)
+    PointLight(Point3 center, double intensity, double radius)
     {
         this->center = center;
         this->intensity = intensity;
